Python_Code/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
types-psycopg2==2.9.10
pandas==2.2.3
pandas-stubs==2.2.3.250308
requests==2.32.3
types_requests==2.32.0.20250328
